+++ client/app/app.js
@@ -95,9 +95,10 @@ angular
         return $sanitize(translationId);
     })
     .config((LANGUAGES, $translateProvider, constants) => {
-        let defaultLanguage = "fr_FR";
+        let defaultLanguage = constants.DEFAULT_LANGUAGE;
 
-        if (localStorage["univers-selected-language"]) {
+        // if there is a stored language value, be sure it's a valid one
+        if (localStorage["univers-selected-language"] && _.find(LANGUAGES, { value: localStorage["univers-selected-language"] })) {
             defaultLanguage = localStorage["univers-selected-language"];
         } else {
             localStorage["univers-selected-language"] = defaultLanguage;
@@ -109,11 +110,11 @@ angular
 
         $translateProvider.useMissingTranslationHandler("translateMissingTranslationHandler");
         $translateProvider.useLoaderCache(true);
-        $translateProvider.useSanitizeValueStrategy("escapeParameters");
+        $translateProvider.useSanitizeValueStrategy("sanitizeParameters");
 
         $translateProvider.preferredLanguage(defaultLanguage);
         $translateProvider.use(defaultLanguage);
-        $translateProvider.fallbackLanguage("fr_FR");
+        $translateProvider.fallbackLanguage(constants.FALLBACK_LANGUAGE);
     })
     .config(($transitionsProvider, $httpProvider) => {
         $httpProvider.interceptors.push("translateInterceptor");
@@ -125,6 +126,6 @@ angular
         OtrsPopupProvider.setBaseUrlTickets(_.get(constants, "REDIRECT_URLS.listTicket", null));
     })
     .run(($translate) => {
-        moment.locale($translate.use().split("_")[0]);
+        moment.locale(_.first($translate.use().split("_")));
     })
     .constant("UNIVERSE", "DEDICATED");